Select active Pokemon
Added ability to select active pokemon at the beginning of the game.
This actually raised a question about control. Messages from the server
to the client are now handled in a much better way.

There is one final issue to fix - the players should be able to
concurrently select their active pokemon.

Signed-off-by: David Isenor <d.isenor@dal.ca>