#16 Refactor Controller and Model
Deprecated several methods in controller and model. Moved functionality
from controller to model, including the playerTurn variable, which is
now controller through the turn variable of the model. Added several
useful functions to the model.

Also added new variable 'knockedOut' to Pokemon class. This was to set
the framework for input from the non-turn player that is necessary for
continuation of play, i.e. getting your active Pokemon knocked out.

Signed-off-by: David Isenor <d.isenor@dal.ca>